Bug#35750771 Refactor Transporter and NdbSockets to handle lifecycle issues
Disconnect of transporters were not 'clean':

disconnectImpl() is used to DISCONNECTING the connection, while other threads
might still send/recv on the same connection. The implementation of disconnectImpl()
used to call NdbSocket::close() directly, such that:

 - close() was called on the underlying socket, which also released the descriptor
   for reuse. As the descriptor was still referred by other threads, they may
   end up referring a descriptor reused to referring e.g. a file.

 - With introduction of TLS in NdbSockets, close() released the ssl
   object as well. The ssl object may also be referred by other threads
   attempting send/recv.

Patch avoid releasing these objects as part of the DISCONNECTING phase:

When we disconnect (-> ::disconnectImpl()) a transportrer, we will now
only 'shutdown()' the NdbSocket which keep the file discriptor and ssl object
in a 'shutdown' state

When the transporter reach the DISCONNECTED state (-> report_disconnect()) we
will call the new Transporter method to releaseAfterDisconnect() which will
release these objects as well.

Lifecycle of the NdbSocket::mutex is also changed, such that it is not
released by a 'close' any longer. It will exist until NbdSocket is
destructed. The Guard2's used with mutex are changed to use Guard instead,
which require the mutex to exists. -> Guard2 just skipped the locking if
there were no mutex.

-> The NdbSocket::enable_locking() and ::disable_locking() methods became obsolete.

Other related and required changes:

    - Fixed testSecureSocket.cpp:
      -- build break.
      -- Added missing stopSession(), needed to close the NdbSocket.
      -- Removed 'locking' argument to several c'tors and methods
         (Locking is now implicit if SSL is used)
      -- Had to disable a test case using SSL (locking) and 'blocking' communication.
         -> Need to discuss, do not think this combination can ever work.

    - Introduced missing 'lifecycle handling' in Loopback and SHM_Transporter.
      -- Introduced Transporter::disconnectImpl() and ::releaseAfterDisconnect()
         implemennting the 'lifecycle' of 'theSocket'
      -- Let TCP, SHM and Loopback_Transporter implementation use the above methods.

    - Removed some more Transporter locks:
      -- SHM and Loopback will not need lock in disconnectImpl() when implemented
         as only a 'shutdown', ... assuming that disconnecting protocol is followed
         (Multi-transporters breaks it ...!)
      -- During connection setup the transporter is not yet public available,
         thus locks should not be needed

   - EPOLL handling can now receive a 'EPOLLHUP' event when from a socket
     being shutdown. Handling of that event need to 'unlisten' to events on
     that socket.

Note that the disconnect of a MultiTransporter does not follow the disconnect
'protocol'. The transporters being 'part of' the MultiTransporter was closed
directly when the node communication itself has reached the DISCONNECTED state.

At this stage the MultiTransporter has attempted to stop communication on the
underlying transporters. However it is propably not guaranteed, and the two-step
disconnecting-and-shutdown -> disconnected-and-close protocol *is* not followed.

As a temporary workaround this patch introduce Transporter::forceUnsafeDisconnect()
in order to make it clear where MultiTransporters are disconnected without following
this protocol. Implementation of that method is similar to the 'old' doDisconnect()
where we closed and released the transporter resources without waiting for the
DISCONNECTED state to be reached.

Bug 35750394 'MultiTransporter should follow the DISCONNECT protocol' is intended to
fix this longer term.

Change-Id: I19d201ddd42596da6eac071f444bdae367586b65