Bug#35515899 connection abort while statement executes does not abort statement
When the client executes a long running statement like
"SELECT SLEEP(100)" through router and aborts the client,
the statement will keep running until a resultset is sent.

This differs from executing the statement directly on the server,
where a SLEEP() is aborted after ~5 seconds.

Root Cause
==========

When the router as sent a statement to the server for executing, it
waits for the servers response and does not check monitor the
client side.

The connection close of the client is only noticed when router sends the
servers response.

Change
======

- check the client-socket readability when waiting for the query's
  response.
- refactored the "watch client and server for readablity" from the
  existing classic_command.cc "server closes connection while waiting
  for the client's next command" into await_client_or_server.cc
- refactored the forward-tls for the PASSTHROUGH handling to use the
  same await_client_or_server.cc

Change-Id: I9c8660d1d8702a2977195813b81e1449b1b092a1