Bug#34846823 CREATE FULLTEXT INDEX fails with ERROR 1062 (23000): Duplicate entry 'NULL-NULL'
Bug affecting fulltext indices spanning multiple columns. Happens when
document queue underruns between processing indexed fields belonging to
the same row (document), which is random and time-dependent. For the
purpose of token offset calculation, the subsequent field gets treated
as a new document, which causes the indexed position to be incorrect.

Change-Id: I77fdab326a13d768024f6eb0e626f89df6a49759