Bug#35714597 Add multi file support to mtr option --skip-test-list
This patch the following feature that eases maintaining multiple
disabled file for the users:

If an user has some common tests that needs to be disabled in the
multiple disabled files then user can define those tests in one file,
and include that file in the other disabled files where ever required
using keyword @import as following :

@import <disabled_file>

<disabled_file> could be just file name or absolute path.
A file is included only once to avoid circular dependencies and
parsing the files multiple times.

File name could be an absolute path or relative to known directories
where these disabled files are usually located.
File is looked at the following location in the order listed:

- <build_dir>/mysql-test, if MTR_BINDIR env variable is set
- <source_dir>/mysql-test/collections

Change-Id: I985fa1f0308235659b5b544c8cc54d934bc1584b