BUG#35749521 receieve_event does not handle new columns
Problem:
Event subscription on a NDB table fails to handle the case when data
changes occur on a subscribed table which is altered inplace to add a
new column at the end.

Analysis:
The NdbEventOperation subscribes to the NDB table and builds a list of
its columns. When the subscribed NDB table data is modified, data change
events are sent to the subscriber and its column list is used to check
if each received column is primary key.

After altering the NDB table to have a new column, subsequent data
change events of the NDB table will contain data also for the new column
(this is by design). The handling of data change events for these newly
added column fails to find the column in the list of subscribed columns
and thus causes an abnormal termination when checking if the column is
primary key.

Solution:
Since the additional data for newly added column is sent from NDB by
design the problem is fixed by handling data columns which are not
in the list of subscribed columns as if it was not primary key.

Change-Id: I08fbbc43e9f83e63ab69d7a96d88e46ef164b752