Bug#35760551 MysqlClient::prepare() should take a std::string_view
MysqlClient::prepare() wraps mysql_stmt_prepare() which takes a
pointer to the start of a string and length.

Currently, a std::string is used as input-argument, which creates a
temporary copy if the passed argument is a c-string.

As mysql_stmt_prepare() takes a pointer and a length, a std::string_view
would avoid the copy and achieve the same.

Change
======

- changed MysqlCli::prepare() from std::string to std::string_view

Change-Id: I969c39ce32bc1d01ad910ca93650f5c67d6eba76