WL#15738: Flush Mask Dictionary from table to memory
All components that have tables that can be changed on a primary and run
on replicas - need to include the ability to call the scheduler
component and flush the data on the secondary or replica into memory.

This worklog is to implement the ability to periodically refresh the
in-memory cache of masking dictionaries from the underlying table.

The flush may be requested by the user ad hoc (by calling UDF) or
periodically, leveraging the scheduler component. In particular, the
periodical flush is required in MDS scenarios.

Change-Id: Ib57b4f5abdd404e39fb89fd3348245d10f4296c1