Bug#35640263: Refactoring of json_diff/json_binary to enable HW support for partial json [3/6, error handler, noclose]
The functions that serialize JSON values currently take one error
handler argument for each error that could be raised by the function.
This does not scale well, as one new argument has to be added for each
new error condition that is added to the function, and many function
signatures have to be modified if errors are added or removed from one
of these functions.

This patch replaces the many error handler arguments with a single
error handler that has one virtual member function per error. This
allows new errors to be added without changing the signatures of the
serialization functions.

Change-Id: If341dbcb98e556c1341934cb69f4427aafe9708c