Bug#35753492 Warn instead of error if not finding expected DD properties
WL#15697 added new DD properties that are expected to be present
in 8.2+. This does not hold for data directories that already
exist for test purposes, and makes the restart fail.

This patch will relax the check for the expected properties and
issue a warning if they are not present. Instead of a failure,
the server will start using default values for the properties.

Change-Id: I65a6ee97d35acd4fde982d68ed060a5b8919b5be