Bug#35730982: GROUP_CONCAT results were displayed as NULL before 8.0.23
This is a followup patch to handle a failing test in RAPID.

The first patch added code that evaluated a const-for-execution
expression in Item_func_group_concat::setup(). However, this might
attempt to evaluate a subquery in RAPID, which is not possible.
To fix this, we add another check evaluate_during_optimization()
which rejects subqueries, among others.

Change-Id: Ideaa3a138fe039586e2d7d18272552eccd2f322c