Bug#35640263: Refactoring of json_diff/json_binary to enable HW support for partial json [4/6, serialize, noclose]
Some error checking and error reporting in the json_binary::serialize*
functions is wrapped in "#ifdef MYSQL_SERVER" so that it is not
invoked when called from client code. This does however not allow them
to be called from plugin code running in the server without a THD.

This patch moves the remaining error checking and reporting into the
error handler, so that the functions can also be called from
server-side code that doesn't have a THD, if an appropriate error
handler object is provided. Specifically, the code for checking and
reporting stack overrun and internal errors is stubbed out.

Change-Id: Ic4841121e313a1c763658388e2ef1ad4e0c6871e