Bug#35621842: Rollup subquery segv
The problem here is that a nameless field is presented to the function
find_item_in_list(). Since the search is name-based, it will fail.

The culprit is in Item::split_sum_func2(), which when presented with
an expression that contains a GROUPING field is supposed to split
the function's arguments, however it adds a reference to the expression.

The function is enhanced so that it also checks for this case.

In addition, we perform some simple refactoring of functions
Item::split_sum_func() and Item::split_sum_func2().

- A standard bool return value is added, so that we can track errors
  in operation without having to check THD::is_error().

For Item::split_sum_func2(), logic and description is improved.

- Terminology is replaced with SQL standard terms, ie. set function,
  grouping function, aggregate function.

- Overall description is rewritten for clarity.

- Different types of items are clearly identified (see variables
  is_aggr_func, is_grouping and (the existing) m_is_window_function.

- Structure of complicated if tests is slightly improved.

- The decision on whether to call Item::split_sum_func() or not is
  generalized. Now it is called for any item that is not itself an
  aggregate function, a grouping function or a windows function, but
  contains one (or more) of these.

- The decision on whether to replace the item with a reference is
  generalized to check for any non-constant item, but with certain
  exceptions (these exceptions are not changed).

Change-Id: I06e266ca5e14595ef875fbb05e5ada813b620bdf