Bug#35757024 Do not log TLS info to stdout in NDBAPI utilities
TlsKeyManager::init() writes a log message to the global EventLogger
including the path name to the certificate file. Previously, this
message was written for DB, MGMD, and API nodes.

The problem was that for many misc. utilities, this log message was
disruptively printed to the standard output or standard error stream.

With this change, the message is only written for DB and MGMD nodes.
Now an API that node desires such a log message must produce the
message itself.

Creating such a log message requires a new public NDB API method:

  Ndb_cluster_connection::get_tls_certificate_path()

This patch implements the new method in Ndb_cluster_connection and down
through the layers of TransporterFacade, TransporterRegistry, and
TlsKeyManager. It also adds code to the ndbcluster plugin to call into
the new method and generate a log message for the MySQL server.

Change-Id: I278c81657c9c923cae3ab0528557ee2b91ff08ce