Bug#34976138 records_in_range does too many disk reads
Index dive optimization logic implemented as part of WL#6526 does not
cover the scenario where an ORDER BY is present and the given index
can provide the required order. This leads to even simple queries with
ORDER BY performs extra disk access and degrade performance.

A new logic is added to the implementation to skip index dive when
ORDER BY is present and the given FORCE INDEX can provide the
required order. Prior to this commit, we would never skip index dives
when doing 'explain', as we wanted better row estimates in 'explain'
output. But this meant that for some queries, 'explain' would
show a query plan using other access methods than those chosen
when planning the same query without 'explain'. For that reason,
we no longer check for 'explain' when deciding whether to skip
index dives.

Change-Id: I312a7330e1074ce7d012231c9a0753a275ab508c