WL#15811: [InnoDB Changes] HeatWave Statically Guided Load
This patch introduces a new parameter 'max_desired_threads' in the
parallel_scan_init() function signature, which is needed as a
prerequisite for the Statically Guided Load worklog.

'max_desired_threads' allows callers of parallel_scan_init() to directly
limit the number of read threads they need, instead of relying on
something external to specify that limit. Passing 0 for this argument
has no effect.

Specifically in the InnoDB plugin, the read thread limit is controlled
by the 'innodb_parallel_read_threads' variable. With this patch, the
number of read threads can be set lower than the current value of the
session variable by providing a value to 'max_desired_threads' argument
of parallel_scan_init().

An MTR test in the rapid suite is added to test this behavior.

Change-Id: Ida397c0ef9cd505186cb7006b1f180688bfb587d