WL#15811: HeatWave Statically Guided Load
This worklog improves the user experience for loading tables to HeatWave
by automating preparatory work necessary to load tables in an optimal
manner.

This worklog aims to provide optimal load parameters before the load
starts. It does not address cases where a load failed due to e.g., out
of codes error, out of memory, too wide columns, etc. Subsequent
worklogs in the load improvement area will tackle these issues.

We use Autopilot's load recommendations to improve the load experience
in four ways:
 1. Setting SECONDARY_ENGINE=RAPID,
 2. Removing unsupported columns before the load starts,
 3. Reducing likelihood of OOMs during load,
 4. Prevent loading of tables that cannot fit into the cluster
    [Lakehouse only].

~~~ 1. Setting SECONDARY_ENGINE=RAPID ~~~

If a secondary load is attempted on a table without a secondary engine,
the following error is reported:
    "Secondary engine operation failed. No secondary engine defined."

Provided that the HeatWave plugin is installed, and the secondary engine
of a table is NULL, we will automatically set the secondary engine to
RAPID, eliminating one step in the load process.

~~~ 2. Removing unsupported columns before the load starts ~~~

Some columns are currently not supported by HeatWave, and have to be
excluded before the load starts. For example, trying to load a table
with a BLOB column will report:
    "Data type blob not supported by RAPID"

Since we know which columns are not supported before the load starts, we
can exclude these columns from secondary load by marking them with the
"NOT SECONDARY" comment.

~~~  3. Reducing likelihood of OOMs during load ~~~

Users can frequently encounter OOMs during load if the
innodb_parallel_read_threads is not correctly configured and determining
an optimal number of read threads helps in reducing OOMs.

Setting this load parameter is slightly different than the previous two,
since Autopilot suggests changing the *session*
innodb_parallel_read_threads variable. Since Guided Load runs from a
different session than the user-issued secondary load, executing that
command in the Guided Load session has no effect. Therefore, we have to
store the recommended value in rpd_mirror, and apply it from the user
session.

~~~ 4. Prevent loading of tables that cannot fit into the cluster ~~~
~~~                       [Lakehouse only]                        ~~~

For Lakehouse it is very important to know whether a table can fit into
the cluster before the load starts, as the entire transformation to
HeatWave format needs to happen before the load to HeatWave nodes is
attempted. This means that loading a 100TB table into a 32GB cluster
might take hours before the cluster goes OOM.

This worklog aims to prevent these situations by rejecting loads that
cannot fit into the cluster. To do this we are relying on Autopilot
table size estimates and simple logic to check whether the table can fit
into an empty cluster:

 estimated_table_size < BASE_MEMORY_QUOTA * total_cluster_size,

where BASE_MEMORY_QUOTA is the percentage of rpdserver memory reserved
for base table relations.

Note that this worklog can still provide false positives and false
negatives, depending on the accuracy of table size estimates. Later
patches can easily tweak this check to provide a better accuracy rate.

The load of the table can be attempted without memory checks by calling
Auto Parallel Load and providing the "'use_mem_checks': FALSE" argument.

~~~~~ Other topics ~~~~~

~~~ Disclaimer ~~~

This worklog relies on Autopilot predictions for estimating the optimal
number of innodb_parallel_read_threads. OOMs during secondary load can
still occur, regardless of the number of used parallel read threads.

~~~ Reporting ~~~

The user can inspect the workings of Guided Load in several ways:

    1. Calling 'SHOW CREATE TABLE <table>;'

    The alteration of a secondary engine, or the exclusion of an
    unsupported column can be expected by running the SHOW CREATE TABLE
    statement.

    2. Calling 'show warnings'

    When the secondary load finishes, all the modifications to the table
    by Guided Load will be reported as warnings. Example:

+---------+------+-----------------------------------------------------+
| Level   | Code | Message                                             |
+---------+------+-----------------------------------------------------+
| Warning | 3877 | Guided Load executed 'ALTER TABLE `S1`.`t1` MODIFY  |
|         |      | `b` blob DEFAULT NULL COMMENT '' NOT SECONDARY'.    |
| Warning | 3877 | Guided Load executed 'ALTER TABLE `S1`.`t1`         |
|         |      | SECONDARY_ENGINE=RAPID'.                            |
| Warning | 3877 | Guided load used 1 InnoDB parallel read threads to  |
|         |      | load S1.t1 (session variable value: 4).             |
+---------+------+-----------------------------------------------------+

    3. Inspecting performance_schema.rpd_mirror.

    The recommended number of read threads can be seen by checking the
    RECOMMENDED_READ_THREADS column from performance_schema.rpd_mirror.
    Note that this value will be NULL if Guided Load was not called for
    this table. Additionally, the value is only meant to serve as a
    guideline, and will not change dynamically based on newly
    added/removed data in a given table. The value will be updated the
    next time Guided Load is called for that table.

~~~ Debugging ~~~

In order to verify that Autopilot runs successfully, we are dumping the
content from sys.heatwave_autopilot_report once the preparatory
statements are executed.

~~~ Reload from InnoDB after restart ~~~

This worklog also reduces the chance of encountering OOM during reload
of a table from InnoDB after system crash or restart. The same Autopilot
recommendations for setting innodb_parallel_read_threads are used as in
the regular user load flow. Note that Guided Load will not execute DDLs
to modify the table structure; only the innodb_parallel_read_threads
will be set.

Change-Id: I51e304c6a28ca337f31d5678a82f85ff1b654374