BUG#35397276: Relax GR member join version compatibility rules
Currently in Group Replication, a server can only join a group if
the server version is equal or higher than the lowest group member.
Example, on a group with S1 (8.0.32), S2 (8.0.33) and S3 (8.0.33),
to be allowed to join, the joining member must be 8.0.32 or higher.

With the new release model supporting LTS and innovation releases,
inplace downgrade should be supported, but only within one LTS
release series.
For server versions 8.0.35+ (8.0.36, 8.0.37, etc.), which are not
strictly LTS releases, inplace downgrade shall still be allowed.

The same reasoning applies to cross-version groups, that is, a
server 8.0.35+ should be allowed to join a 8.0.36 or higher 8.0
patch version.
Example, on a group with S1 (8.0.36) and S2 (8.0.37), a server which
version is 8.0.35 is allowed to join.

For this rule to apply all servers version must be 8.0, that is,
if a higher major version server is part of the group it does not
apply.
Example, on a group with S1 (8.0.36) and S2 (8.1.0), a server which
version is 8.0.35 is not allowed to join.

The change of this rule will not impact the primary election
algorithm.
On a primary failover, the elected primary will be
from the lowest version.
Likewise, only the lowest version servers can be used on
group_replication_set_as_primary().

Change-Id: I80dc66e756c837f5dbe0bf0e9d3177ccfc8bbb1e