Bug#35240825 performance_schema.data_locks query increased RAM usage to OOM
Problem
=======

1)

For a server in production, under load, the query:

SELECT * FROM performance_schema.data_locks;

can consume a lot of memory,
possibly causing the server to fail.

2)

The memory allocation done in table performance_schema.data_locks
is not instrumented.

Root cause
==========

1)

The original implementation of table performance_schema.data_locks
uses an internal cache, PFS_data_lock_container,
to hold data reported by the innodb storage engine.

While the PFS_data_cache itself works properly,
an extra copy of the data is present in the rows,
embedded in row_data_lock.

This is because fragments like row_data_lock::m_index_row
do have an extra copy of the table schema and name,
in nested structures PFS_schema_name and PFS_object_name.

Also, the row_data_lock structure holds space for the index name,
even when no indexes are used.

Keeping a copy of schema and table name defeats the purpose
of sharing these names in a cache in the first place,
and leads to a huge memory consumption.

2)

PFS_data_lock_container uses C++ STL containers,
which uses the default memory allocator from the STL.

The performance schema memory instrumentation is missing,
so that allocations are not observable.

Fix
===

1)

The row structure row_data_lock has been revisited.

Classes like PFS_schema_name_view are implemented.

PFS_schema_name_view behaves the same way as PFS_schema_name,
except that it does not own the data,
so that several rows having the same value in the schema name column
will effectively points to the same memory (in PFS_data_cache)
that holds the schema name value.

Likewise for PFS_object_name_view, to represent a table name.

Likewise, class PFS_object_name_view was created,
to optionally point to a shared index name in the cache.

Now, because PFS_schema_name_view, PFS_object_name_view classes
to share existing memory,
these classes can no longer internally reformat a name to comply
with the lower_case_tale_name rules.

When a PFS_schema_name_view object is created using a given name,
the name must be normalized already.

A consequence is that names must be normalized before adding
them in the PFS_data_cache.

To help with this, the interface exposed to the innodb engine
has been extended with cache_itentifier(),
where the caller (innodb) indicate what an indentifier represent,
so that in turn the data container can apply the proper formatting
rules before adding an identifier in the data cache.

2)

A class, PFS_std_allocator<T>, provides memory instrumentation,
and can be used as allocator when using the C++ STL.

Class PFS_data_container_allocator is an allocator bound to
a new internal memory instrument, "memory/performance_schema/data_container".

Containers are now used with the instrumented allocator,
and report usage statistics in "memory/performance_schema/data_container".

3)

During investigations, it was also found by code review that
the primary key for:
- table performance_schema.data_locks
- table performance_schema.data_lock_waits

was not safe.

The hidden primary key structures:
- pk_pos_data_lock
- pk_pos_data_lock_wait

are used with memcmp() operations,
and every byte in the key should be set when writing to it.

However, when the data container reused internal rows
from a scan to the next, an existing pk structure was reused with
a new key value.

In this case, a new key was written with a memcpy() in:
- PFS_data_lock_container::add_lock_row()
- PFS_data_lock_wait_container::add_lock_wait_row()
but the code forgot to memset the rest of the key.

This is fixed by refactoring the primary key structures,
and in particular declare the internal buffers private,
forcing the code to use the proper set() method.

Change-Id: Ic2188b58ac3a71d63cb8193142962ae5bacf730c