Bug#35676721 fil_tablespace_redo_rename(unsigned char* recv_parse_or_apply_log_rec_body(mlog_ [noclose]
Add a missing check in fil_tablespace_redo_rename.
While processing the log buffer, a prematurely cut off
buffer content could result in reading uninitialized bytes.
The code expects two bytes in the buffer indicating a 16 bit integer,
but there was no check to verify that those two bytes are actually
in the buffer (as opposed to the buffer ending with fewer than
two bytes left before the end).

More specifically: the routine fil_tablespace_redo_rename
expects to find two C strings, and their length (lenghts including null
terminator chars) in the redo log. If redo log is corrupted, and
ends right at the end of the first string (the filename to be
renamed from), fil_tablespace_redo_rename attempts to continue
processing the redo log, to read the length of the next C string
(the filename to be renamed to). This unconditional read of a 16 bit
integer can result in undefinied behaviour, if the redo log actually
doesn't contain these 16 bits.

Additionally, similar issue are fixed in fil_tablespace_redo_create
and fil_tablespace_redo_delete as well.

To make such bugs easier to detect, we mark all the allocated
memory after the end of the buffer as poisoned in ASAN builds.
Those memory locations might have been initialized before, and
the allocated memory reused for smaller content, hence the memory
following the content was not always marked as uninitialized
in Address Sanitizer builds.

Change-Id: I73760e2ac2df8e81ec6ba306f70206cf32638eeb