Bug#35676721 fil_tablespace_redo_rename(unsigned char* recv_parse_or_apply_log_rec_body(mlog_
Clean up const correctness of buffer pointers in redo log parsing.
This makes it a little easier to reason about the code.
The pointer pointing to bytes in the buffer containing redo log sections
always points to const byte.
This was not consistently followed, which as a side effect resulted in
having a couple of undesirable casts, of the kind:

return (const_cast<byte *>(ptr));

Change-Id: I960b036fb3f4806f9f731d084351a4b2e087788b