Bug#35691447 rw-splitting routes to server removed from group
Given an InnoDB Cluster and a MySQL Router using read-write splitting:

If the primary server is being removed from the cluster group a
read-only operation may be routed to the removed server.

Root-Cause
==========

When the cluster changes its state (nodes are removed, added, ...) it
calls:

   ConnectionContainer::disconnect()

It is passed a positive-list of all destinations which are still
allowed.

With read-write splitting each connection has two destinations:

- a read-write destination
- a read-only destination

The disconnect() function only checked the "current" destination
(read-write XOR read-only).

If the current destination was "read-only", and the "read-write"
destination vanished, the connection was not dropped.

Change
======

- changed ConnectionContainer::disconnect() to drop a connection
  if at least one of the read-only and read-write destinations
  isn't allowed anymore.

Change-Id: I774b5a8e6ef0b83de5a9292bb63ab7a4a77bb0fd