Bug#35755396: audit_plugin: The chrono::steady_clock::now() usage causes performance drop
Fixed a bug in the scheduler logic:
The mysql conditional variable timed wait is expecting an
*absolute* time and the wrapper was giving it a relative time.
This was causing the scheduler background loop to not wait at all
and busyloop check the queue.
Fixed by:
- making sure the mysql conditional variable timed wait
gets an absolute time and also fixed the unit test mock code
to process the time passed to it as an absolute time.
- Replaced the usage of steady_clock::now with
system_clock::from_time_t(time(nullptr)). This cuts down on the
timer resolution to 1sec, but is faster.
- added a missing file in one of the guid unit tests on windows
causing it to not be compiling on windows.
- added one extra second to the wait for the top task to avoid
some limited (to 1 sec) busy-looping on linux when executing the
task.
Addressed review comments.

Change-Id: Id6cdd974384c4132470af7036e8920e50cbcecdc