Bug#35774109 Reset connection m_error_states before (re-) connecting
Ensure that a previous error is cleared before we try to reconnect
a (failed?) connection.

TransporterRegistry::do_connect() is where we initiate a connection
by setting the CONNECTING state. Note that the state is required
to be DISCONNECTED at that point.

Patch sets m_error_states[] for the CONNECTING node to 'NO_ERROR'
as part of validating and initiating CONNECTING.

Change-Id: I143e71fd99da177e8f9afa7f42c40d4b3c400210