Bug#35082022 - Plaintext is displayed in errorlog if process crashes when changing password
When a crash happens on the server, we display stack trace and the
query which caused the crash in the server error logs.

Problem:
--------
If a crash happens while creating or altering a user with a specified
password, the password specified in client-side is displayed in plain
text to the server error logs.

Fix:
----
Use thd->rewritten_query() instead of thd->query() to display the
query, which prints the password in the encrypted form.

Change-Id: I3a21e055a843de0f48f97efc56566b778a761451