Bug#35512099: hypergraph_optimizer SEGV
Some INSERT statements that made use of table value constructors
failed when running in the hypergraph optimizer.

The reason was that the simplified preparation of table value
constructors in Query_block::prepare() skipped a call to
Query_block::setup_tables(), which made Query_block::leaf_tables
contain the wrong set of tables (it would contain the target table of
the INSERT statement instead of being empty). This again made
JOIN::optimize() fail to take the shortcut for table-less queries, so
that the hypergraph optimizer would be called to optimize the query.
Since the hypergraph optimizer does not expect to be called on table
value constructors, it handled it incorrectly and produced an invalid
query plan.

The problem with an incorrect table list in the query block also
affects the old optimizer. The consequences are less severe, though.
It fails to take the shortcut for table-less queries, so it performs
some unnecessary work, but it still ends up producing a valid query
plan.

The fix is to make Query_block::prepare_values() invoke
setup_tables(). This makes the query run successfully in the
hypergraph optimizer, and avoids the unnecessary work in the old
optimizer.

Change-Id: I5f1d2ccdf9f72fccbd7eb667bc1e26464bae5e5e