Bug#35767731 Replace all use of 'utf8' and '_utf8' [.inc noclose]
utf8 is currently an alias for utf8mb3. We want to change this some
time in the future.  Replace all use of 'utf8' and '_utf8' and
'collate utf8_xxx' in .inc files, and re-record .result files.

The patch does the same replacement for the 'funcs_1' test suite.

It also re-enables main.internal_tmp_disk_storage_engine:
Disable a couple of updates, which previously failed.
Replace int(11) with int to avoid 'Integer display width' warnings.
This is to get a smaller diff.

Skip the entire test for hypergraph, and use format=tree rather than
json. Cost estimates are not stable, and we have --elide support for
format=tree only.

Change-Id: Ibb1d58485efbf4639aa0a2a517b97e8a19ad1bc7