Bug#30775369: WRONG RESULT WITH "IN" COMPARATOR AND A TABLE VALUE CONSTRUCTOR
If a table value constructor was used as an IN subquery, and the IN
subquery was transformed to an EXISTS subquery, the resulting query
plan did not have a filter for the WHERE clause generated by the
IN-to-EXISTS transformation. Too many rows were returned from the
query.

Fixed by adding a filter on top of the table value constructor if it
has a synthetic WHERE clause.

Change-Id: I6e97771cc015f24dabcfc7a515ef40a0574284c5