Bug#35736158: Insert on duplicate assertion error
An inconsistency is detected when resolving a subquery in the
expression list of the ON DUPLICATE KEY UPDATE part of an INSERT
statement. Such expressions are attached to the first query block of
the INSERT, which is essential if the query expression of the INSERT is
a set operation (e.g. a UNION). However, these expressions are instead
attached to the last query block of the set operation.

The problem is in contextualization where the query block pointer points
to the wrong query block in PT_insert::make_cmd() just before
contextualizing the ODKU expressions. The fix is to restore this pointer
in finalize_query_expression(). Thus, we also no longer need to save
current query block in callers of finalize_query_expression().

Change-Id: Idb675ccef9eec1fc7d4100b5fd709cccae309fb4