Bug#35719688 Assertion`new_path->cost >= new_path->init_cost' in ExpandSingleFilterAccessPath
This bugs is partly a regression due to the commit with Change-Id:
I34c307b12216025bedf6e995ee89a58793f59aac (fix for bug#34787357),
and partly a consequence of a preexisting error in init_cost
calculation for ROWID_UNION.

This bug is triggered by a floating-point rounding error when
calculating the init_cost of a ROWID_UNION access path. This rounding
error becomes significant when the relative difference between
'init_cost' and 'cost_before_filter' is small. Depending on which way
the rounding goes, this may cause the 'cost >= init_cost' assert to
fail.

For this to happen, 'init_cost' and 'cost' must be identical (or very
close). To fix this, two things are changed:

- init_cost for ROWID_UNION is set to the sum of the init_cost of
  its child scans. Prior to this fix, it was set equal to 'cost',
  which is wrong.

- Item_func_ne::get_filtering_effect() should not return zero
  selectivity. Returning zero was a necessary condition for the
  assert to fail (as 'cost' otherwise would be greater than
  'init_cost'). Also, returning zero may give distorted cost figures,
  as e.g. the inner side of a nested-loop join appears to be 'free' if
  the outer side return zero rows. So if we return a small non-zero
  value, there is less chance of making very poor plan choices.

Note that init_once_cost for ROWID_UNION is now calculated as the sum
of the init_once_cost of the child scans. Prior to this fix, it was
set equal to 'cost'. This had the side effect of making
rowid_union_path->rescan_cost()==0, which again meant that nested loop
joins with a ROWID_UNION on the 'inner' side became artificially
cheap.

Change-Id: Ieaee80d18e27ca937b1776d6ff72484a5f62a251